fix_feature23/put decorator in utils.py
